Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src/script-src-1_4_1-expected.txt

CONSOLE MESSAGE: Refused to execute a script because 'unsafe-eval' does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to execute a script because 'unsafe-eval' does not appear in the script-src directive of the Content Security Policy.
setTimeout() and setInterval() should not run without 'unsafe-eval' script-src directive.


Harness Error (TIMEOUT), message = null

PASS window.setTimeout()
PASS window.setInterval()
NOTRUN Test that securitypolicyviolation event is fired
